// src/types/cart.ts
export type CartItem = {
  productId: string;
  variantId: string;
  productSlugFr: string;
  productSlugEn: string;
  productNameFr: string;
  productNameEn: string;
  variantLabelFr: string;
  variantLabelEn: string;
  imageUrl: string;
  unitPriceMillimes: number;
  quantity: number;
  maxStock: number;
};